requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1

# Flask
Flask==1.1.2

# Google Auth
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0

# Google API Client
google-api-python-client==2.32.0

# Requests
requests==2.26.0

# pathlib (included in Python standard library)
pathlib2==2.3.6

# cryptography
cryptography==36.0.1

# pip-vendor (included in Python standard library)
pip-vendor==0.3.1
